feature : CQRS 패턴에 맞게 code단에서 Class 변경 (TokenGenerator -> TokenCommandServer, TokenManager -> TokenQueryService) 및 그에 맞게 메소드 이동(TokenManager에 있던 provideShare를 TokenCommanService로 이동
