+++ src/main/java/team/moebius/disposer/controller/DisposerController.java
@@ -14,22 +14,22 @@ import team.moebius.disposer.dto.ReqReceive;
 import team.moebius.disposer.dto.ReqRetrieval;
 import team.moebius.disposer.dto.RespDistribution;
 import team.moebius.disposer.dto.RespReceive;
-import team.moebius.disposer.service.TokenGenerator;
-import team.moebius.disposer.service.TokenManager;
+import team.moebius.disposer.service.TokenCommandService;
+import team.moebius.disposer.service.TokenQueryService;
 import team.moebius.disposer.util.DateTimeSupporter;
 
 @RestController
 @RequiredArgsConstructor
 public class DisposerController {
 
-    private final TokenGenerator tokenGenerator;
-    private final TokenManager tokenManager;
+    private final TokenCommandService tokenCommandService;
+    private final TokenQueryService tokenQueryService;
 
     @PostMapping("/distribute")
     public ResponseEntity<RespDistribution> distribute(@RequestHeader("X-USER-ID") Long userId,
         @RequestHeader("X-ROOM-ID") String roomId, @RequestBody ReqDistribution reqDistribution) {
 
-        String tokenKey = tokenGenerator.generateToken(
+        String tokenKey = tokenCommandService.generateToken(
             userId,
             roomId,
             reqDistribution.getAmount(),
@@ -44,7 +44,7 @@ public class DisposerController {
     public ResponseEntity<RespReceive> receive(@RequestHeader("X-USER-ID") Long userId,
         @RequestHeader("X-ROOM-ID") String roomId, @RequestBody ReqReceive reqReceive){
 
-        long shareAmount = tokenManager.provideShare(
+        long shareAmount = tokenCommandService.provideShare(
             userId,
             roomId,
             reqReceive.getTokenKey(),
@@ -58,7 +58,7 @@ public class DisposerController {
     public ResponseEntity<TokenInfo> RetrieveDistributionInfo(@RequestHeader("X-USER-ID") Long userId,
         @RequestHeader("X-ROOM-ID") String roomId, @RequestBody ReqRetrieval reqRetrieval){
 
-        TokenInfo tokenInfo = tokenManager.provideInfo(
+        TokenInfo tokenInfo = tokenQueryService.provideInfo(
             userId,
             roomId,
             reqRetrieval.getTokenKey(),
