+++ src/main/java/team/moebius/disposer/service/TokenQueryService.java
@@ -10,7 +10,6 @@ import team.moebius.disposer.domain.TokenInfo;
 import team.moebius.disposer.entity.Recipient;
 import team.moebius.disposer.entity.Token;
 import team.moebius.disposer.exception.NotFoundTokenException;
-import team.moebius.disposer.exception.RecipientException;
 import team.moebius.disposer.exception.TokenException;
 import team.moebius.disposer.repo.RecipientRepository;
 import team.moebius.disposer.repo.TokenRepository;
@@ -18,28 +17,16 @@ import team.moebius.disposer.util.DateTimeSupporter;
 
 @Service
 @RequiredArgsConstructor
-public class TokenManager {
+public class TokenQueryService {
 
     private final TokenRepository tokenRepository;
-
     private final RecipientRepository recipientRepository;
-
-    @Transactional
-    public Long provideShare(long userId, String roomId, String tokenKey, Long targetTime) {
-
-        Token token = checkIsPresentAndGetToken(roomId, tokenKey);
-
-        filterDistributorRequest(userId, token, true);
-        checkReceiveExpTime(token, targetTime);
-
-        return findReceivableRecipient(token, userId);
-    }
 
     @Transactional(readOnly = true)
     public TokenInfo provideInfo(long userId, String roomId, String tokenKey, Long targetTime) {
 
         Token token = checkIsPresentAndGetToken(roomId, tokenKey);
-        filterDistributorRequest(userId, token, false);
+        checkIsDistributor(userId, token, false);
         checkReadExpTime(token, targetTime);
 
         List<Recipient> receiveRecipients = getReceiveRecipients(token);
@@ -51,7 +38,7 @@ public class TokenManager {
         );
     }
 
-
+    // 조회 작업에 대한 응답값을 구성 한다.
     private TokenInfo buildTokenInfo(Token token, long totalAmount,
         List<ReceiveInfo> receiveInfos) {
 
@@ -65,6 +52,7 @@ public class TokenManager {
             .build();
     }
 
+    // Token에 해당하는 받기 작업이 가능한
     private List<Recipient> getReceiveRecipients(Token token) {
         return recipientRepository.findReceiveAllByToken(token.getId());
     }
@@ -81,58 +69,14 @@ public class TokenManager {
             .sum();
     }
 
-    private Long findReceivableRecipient(Token token, long userId) {
-
-        List<Recipient> recipients =
-            recipientRepository.findAllByTokenId(token.getId());
-
-        checkAlreadyReceiveUser(recipients, userId);
-
-        Recipient recipient = getReceivableRecipient(recipients);
-        recipient.setUserId(userId);
-
-        return recipient.getAmount();
-    }
-
-    private Recipient getReceivableRecipient(List<Recipient> recipients) {
-
-        Optional<Recipient> optionalRecipient = recipients.stream()
-            .filter(recipient -> recipient.getUserId() == null)
-            .findAny();
-
-        if (optionalRecipient.isEmpty()) {
-            throw new RecipientException("The distribution has been fully received");
-        }
-
-        return optionalRecipient.get();
-    }
-
-    private void checkAlreadyReceiveUser(List<Recipient> recipients, long userId) {
-
-        boolean isAlreadyReceiveUser = recipients.stream()
-            .anyMatch(
-                recipient -> recipient.getUserId() != null && recipient.getUserId() == userId
-            );
-
-        if (isAlreadyReceiveUser) {
-            throw new RecipientException("User has already received a share from the distribution");
-        }
-    }
-
-
-    private void checkReceiveExpTime(Token token, long targetTime) {
-        if (token.getReceiveExp() <= targetTime) {
-            throw new TokenException("The token has expired for receiving.");
-        }
-    }
-
     private void checkReadExpTime(Token token, long targetTime) {
         if (token.getReadExp() <= targetTime) {
             throw new TokenException("The token has expired for reading.");
         }
     }
 
-    private Token checkIsPresentAndGetToken(String roomId, String tokenKey) {
+    private Token checkIsPresentAndGetToken(String roomId, String tokenKey)
+        throws NotFoundTokenException {
 
         Optional<Token> optionalToken =
             tokenRepository.findTokenByRoomIdAndTokenKey(roomId, tokenKey);
@@ -145,11 +89,8 @@ public class TokenManager {
     }
 
 
-    private void filterDistributorRequest(long userId, Token token, boolean isExcludeDistributor) {
-
-        if (isExcludeDistributor && token.isDistributor(userId)) {
-            throw new TokenException("You cannot receive a share from a token you distributed.");
-        }
+    private void checkIsDistributor(long userId, Token token, boolean isExcludeDistributor)
+        throws TokenException {
 
         if (!isExcludeDistributor && !token.isDistributor(userId)) {
             throw new TokenException("You can only query tokens you distributed.");
